# swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Muslim Daily API Server",
    "version": "2.0.0",
    "description": "Production-grade RESTful API providing comprehensive Islamic digital services: Holy Quran with audio recitations & translations, accurate prayer timings & trackers, authenticated Hadith collections, authentic Duas & Azkar, multi-system calendar conversions (Gregorian, Hijri, Ethiopian), fasting & journal logs, community challenges, and developer utilities.",
    "contact": {
      "name": "Muslim Daily Engineering Team",
      "email": "abdutahir.dev@gmail.com"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "/",
      "description": "Current Server Instance"
    },
    {
      "url": "http://localhost:3000",
      "description": "Local Development Server"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Enter your JWT token obtained from `/api/auth/login` or `/api/auth/register`."
      }
    },
    "schemas": {
      "DictionaryEntry": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "dict_001"
          },
          "word_ar": {
            "type": "string",
            "example": "سَلَام"
          },
          "word_ar_clean": {
            "type": "string",
            "example": "سلام"
          },
          "word_am": {
            "type": "string",
            "example": "ሰላም"
          },
          "word_en": {
            "type": "string",
            "example": "Peace / Greeting"
          },
          "transliteration_ar": {
            "type": "string",
            "example": "Salam"
          },
          "transliteration_am": {
            "type": "string",
            "example": "Selam"
          },
          "part_of_speech": {
            "type": "string",
            "example": "noun"
          },
          "category": {
            "type": "string",
            "example": "Faith & Spirituality"
          },
          "root_ar": {
            "type": "string",
            "example": "س ل م"
          },
          "definition_ar": {
            "type": "string",
            "example": "الأمان والطمأنينة والسكينة وخلو النفس والمجتمع من النزاع."
          },
          "definition_am": {
            "type": "string",
            "example": "መረጋጋት፣ ጸጥታና ፀብ አልባነት፤ እንዲሁም የደህንነት ሰላምታ።"
          },
          "definition_en": {
            "type": "string",
            "example": "A state of harmony, safety, tranquility, and freedom from strife."
          },
          "synonyms": {
            "type": "object",
            "properties": {
              "ar": {
                "type": "array",
                "items": {
                  "type": "string"
                }
              },
              "am": {
                "type": "array",
                "items": {
                  "type": "string"
                }
              },
              "en": {
                "type": "array",
                "items": {
                  "type": "string"
                }
              }
            }
          },
          "antonyms": {
            "type": "object",
            "properties": {
              "ar": {
                "type": "array",
                "items": {
                  "type": "string"
                }
              },
              "am": {
                "type": "array",
                "items": {
                  "type": "string"
                }
              },
              "en": {
                "type": "array",
                "items": {
                  "type": "string"
                }
              }
            }
          },
          "examples": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "ar": {
                  "type": "string"
                },
                "am": {
                  "type": "string"
                },
                "en": {
                  "type": "string"
                }
              }
            }
          }
        }
      },
...
      "AsmaulHusnaItem": {
        "type": "object",
        "properties": {
          "number": {
            "type": "integer",
            "example": 1
          },
          "name_ar": {
            "type": "string",
            "example": "الرَّحْمَٰنُ"
          },
          "name_ar_clean": {
            "type": "string",
            "example": "الرحمن"
          },
          "transliteration": {
            "type": "string",
            "example": "Ar-Rahman"
          },
          "translation": {
            "oneOf": [
              {
                "type": "string",
                "example": "The Entirely Merciful"
              },
              {
                "type": "object",
                "properties": {
                  "en": {
                    "type": "string",
                    "example": "The Entirely Merciful"
                  },
                  "am": {
                    "type": "string",
                    "example": "እጅግ በጣም ሩኅሩህ"
                  },
                  "ar": {
                    "type": "string",
                    "example": "ذو الرحمة الشاملة لجميع الخلائق"
                  }
                }
              }
            ]
          },
          "description": {
            "oneOf": [
              {
                "type": "string",
                "example": "The One whose vast mercy encompasses all of creation in this world."
              },
              {
                "type": "object",
                "properties": {
                  "en": {
                    "type": "string",
                    "example": "The One whose vast mercy encompasses all of creation in this world."
                  },
                  "am": {
                    "type": "string",
                    "example": "በዚህ ዓለም ላሉ ፍጥረታት ሁሉ ምህረቱና እዝነቱ እጅግ ሰፊና ሁሉን አቀፍ የሆነው ጌታ።"
                  },
                  "ar": {
                    "type": "string",
                    "example": "الذي وسعت رحمته كل شيء في الدنيا وعمت جميع الخلائق."
                  }
                }
              }
            ]
          },
          "reference": {
            "type": "object",
            "properties": {
              "quran": {
                "type": "string",
                "example": "Surah Al-Fatihah 1:3"
              },
              "surah_number": {
                "type": "integer",
                "example": 1
              },
              "ayah_number": {
                "type": "integer",
                "example": 3
              },
              "ayah_ar": {
                "type": "string",
                "example": "الرَّحْمَٰنِ الرَّحِيمِ"
              },
              "hadith": {
                "type": "string",
                "example": "Sahih al-Bukhari 2736"
              }
            }
          }
        }
      },
...
      "Dua": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "example": 1
          },
          "category_id": {
            "type": "integer",
            "example": 1
          },
          "title_en": {
            "type": "string",
            "example": "Morning Dhikr"
          },
          "title_ar": {
            "type": "string",
            "example": "أذكار الصباح"
          },
          "content_ar": {
            "type": "string",
            "example": "أَصْبَحْنَا وَأَصْبَحَ الْمُلْكُ لِلَّهِ..."
          },
          "content_en": {
            "type": "string",
            "example": "We have reached the morning and unto Allah belongs sovereignty..."
          },
          "transliteration": {
            "type": "string",
            "example": "Asbahna wa asbahal-mulku lillah..."
          },
          "reference": {
            "type": "string",
            "example": "Sahih Muslim"
          }
        }
      },
...
    }
  },
  "paths": {
...
    "/api/hadith/search": {
      "get": {
        "tags": [
          "Hadith"
        ],
        "summary": "Search Hadiths",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Matching Hadiths"
          }
        }
      }
    },
...
    "/api/duas": {
      "get": {
        "tags": [
          "Duas & Azkar"
        ],
        "summary": "List Duas",
        "parameters": [
          {
            "name": "categoryId",
            "in": "query",
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 20
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Array of Duas",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Dua"
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/qamus/search": {
      "get": {
        "tags": [
          "Qamus Quranic Lexicon"
        ],
        "summary": "Search Quranic Lexicon",
        "description": "Search across Arabic headwords, roots, English definitions, transliterations, and meanings with diacritic normalization.",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Search query string (Arabic or English)"
          },
          {
            "name": "section",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "verb",
                "noun",
                "particle"
              ]
            },
            "description": "Optional filter by section"
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 25
            },
            "description": "Maximum results (max 100)"
          }
        ],
        "responses": {
          "200": {
            "description": "Search results with relevance sorting"
          }
        }
      }
    },
...
    "/asmaul-husna": {
      "get": {
        "tags": [
          "Asmaul Husna"
        ],
        "summary": "List the 99 Names of Allah",
        "description": "Returns the 99 Beautiful Names of Allah with transliteration, Arabic script, English/Amharic/Arabic translations, deep theological descriptions, and Quranic/Hadith citations. Supports localization via `lang` and search/pagination.",
        "parameters": [
          {
            "name": "lang",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "all",
                "en",
                "am",
                "ar"
              ],
              "default": "all"
            },
            "description": "Language localization: 'all' (complete multilingual bundle), 'en' (English), 'am' (Amharic / አማርኛ), or 'ar' (Arabic / العربية)"
          },
          {
            "name": "search",
            "in": "query",
            "schema": {
              "type": "string"
            },
            "description": "Search filter across Arabic script, clean transliteration, English, and Amharic text"
          },
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 1
            },
            "description": "Page number for pagination"
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer"
            },
            "description": "Items per page (1-100). If omitted, all 99 names are returned."
          }
        ],
        "responses": {
          "200": {
            "description": "Array of Asmaul Husna items",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "total": {
                      "type": "integer",
                      "example": 99
                    },
                    "count": {
                      "type": "integer",
                      "example": 99
                    },
                    "language": {
                      "type": "string",
                      "example": "all"
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/AsmaulHusnaItem"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/asmaul-husna/random": {
      "get": {
        "tags": [
          "Asmaul Husna"
        ],
        "summary": "Get a Random Name of Allah",
        "description": "Returns a randomly selected Name of Allah with complete theological explanation, reference, and multilingual translations.",
        "parameters": [
          {
            "name": "lang",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "all",
                "en",
                "am",
                "ar"
              ],
              "default": "all"
            },
            "description": "Language localization"
          }
        ],
        "responses": {
          "200": {
            "description": "Random Name of Allah",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AsmaulHusnaItem"
                }
              }
            }
          }
        }
      }
    },
    "/asmaul-husna/daily": {
      "get": {
        "tags": [
          "Asmaul Husna"
        ],
        "summary": "Get Daily Name of Allah (Name of the Day)",
        "description": "Returns a deterministic Name of the Day based on the calendar date, perfect for daily contemplation and digital widgets.",
        "parameters": [
          {
            "name": "date",
            "in": "query",
            "schema": {
              "type": "string"
            },
            "description": "Optional ISO date (YYYY-MM-DD). Defaults to current day."
          },
          {
            "name": "lang",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "all",
                "en",
                "am",
                "ar"
              ],
              "default": "all"
            },
            "description": "Language localization"
          }
        ],
        "responses": {
          "200": {
            "description": "Daily Name of Allah",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "date": {
                      "type": "string",
                      "example": "2026-09-13"
                    },
                    "name": {
                      "$ref": "#/components/schemas/AsmaulHusnaItem"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/asmaul-husna/search": {
      "get": {
        "tags": [
          "Asmaul Husna"
        ],
        "summary": "Multi-Field Search Across Arabic, Amharic, and English",
        "description": "Executes an instant diacritic-tolerant search across all 99 names in Arabic script, Amharic Ge'ez script, transliterations, theological descriptions, and references.",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Search keyword in Arabic (e.g. 'رحيم'), Amharic (e.g. 'ሩኅሩህ' or 'ንጉሥ'), or English (e.g. 'Merciful')"
          },
          {
            "name": "lang",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "all",
                "en",
                "am",
                "ar"
              ],
              "default": "all"
            },
            "description": "Response language"
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer"
            },
            "description": "Maximum number of results"
          }
        ],
        "responses": {
          "200": {
            "description": "Matching names",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "query": {
                      "type": "string"
                    },
                    "total": {
                      "type": "integer"
                    },
                    "language": {
                      "type": "string"
                    },
                    "results": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/AsmaulHusnaItem"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/dictionary": {
      "get": {
        "tags": [
          "Dictionary"
        ],
        "summary": "List or Search Trilingual Dictionary Entries",
        "description": "Returns paginated dictionary entries with optional search query across Arabic, Amharic, and English, plus filters for category, part of speech, and target language view.",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "schema": {
              "type": "string"
            },
            "description": "Search query across Arabic (with/without tashkeel), Amharic (Ge'ez script), English, transliteration, or root"
          },
          {
            "name": "lang",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "all",
                "ar",
                "am",
                "en"
              ],
              "default": "all"
            },
            "description": "Focus language formatting"
          },
          {
            "name": "category",
            "in": "query",
            "schema": {
              "type": "string"
            },
            "description": "Filter by thematic category (e.g. Faith & Spirituality, Worship & Law)"
          },
          {
            "name": "pos",
            "in": "query",
            "schema": {
              "type": "string"
            },
            "description": "Filter by part of speech (noun, verb, adjective)"
          },
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 1
            },
            "description": "Page number"
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 20
            },
            "description": "Page size"
          }
        ],
        "responses": {
          "200": {
            "description": "List of dictionary records",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "total": {
                      "type": "integer"
                    },
                    "page": {
                      "type": "integer"
                    },
                    "limit": {
                      "type": "integer"
                    },
                    "totalPages": {
                      "type": "integer"
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/DictionaryEntry"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  }
}